Finalize Prototype 2/30/16
Finalized the prototype for class. Made some tweaks. Level txt doc had
to be moved out of JERF, some issues in building due to it being a repo
file

+++ ParentEnemy.cs
@@ -4,7 +4,7 @@ using System.Collections;
 public class ParentEnemy : MonoBehaviour {
 
 	public int hp;
-	protected int fireRate;
+	protected float fireRate;
 	protected int speed;
 	protected float cd;
